api/requirements.txt
Flask
flask-cors
web3
psycopg2-binary
gunicorn
eth-abi
pyjwt
